requirements.txt
asn1crypto==0.24.0
boto3==1.9.168
botocore==1.12.168
certifi==2019.3.9
cffi==1.12.3
cryptography==2.7
docutils==0.14
future==0.17.1
jmespath==0.9.4
pycparser==2.19
python-dateutil==2.8.0
python-telegram-bot==11.1.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
